Initial implementation to fix yubikey field data carryover
