+++ src/v3/src/transformer/uischema/setFocusOnFirstElement.ts
@@ -10,5 +10,25 @@
  * See the License for the specific language governing permissions and limitations under the License.
  */
 
-export * from './setFocusOnFirstElement';
-export * from './transform';
+import {
+  TransformStepFn,
+  UISchemaElement,
+} from '../../types';
+import {
+  isInteractiveType,
+} from '../../util';
+import { traverseLayout } from '../util';
+
+export const setFocusOnFirstElement: TransformStepFn = (formbag) => {
+  let firstFieldFound = false;
+  traverseLayout({
+    layout: formbag.uischema,
+    predicate: (el) => (!firstFieldFound && isInteractiveType(el.type)),
+    callback: (el) => {
+      const uischemaElement = (el as UISchemaElement);
+      uischemaElement.focus = true;
+      firstFieldFound = true;
+    },
+  });
+  return formbag;
+};
